# internal/users/api/dto/users.request.go
package dto

// RegisterRequest represents a registration request
// @Description Registration request payload
type RegisterRequest struct {
	Name     string `json:"name" binding:"required" example:"johndoe"`
	Username string `json:"username" binding:"required" example:"johndoe"`
	Email    string `json:"email" binding:"required" example:"johndoe@gmail.com"`
	Password string `json:"password" binding:"required" example:"xxxxxxx"`
}

type LoginRequest struct {
	Username string `json:"username"`
	Password string `json:"password"`
}